scripts/h2画像生成/requirements.txt
# h2画像生成用
python-dotenv>=1.0
PyYAML>=6.0
Pillow>=10.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
